Add support for gradient stops.
The bilinear interpolation sucks with more than 3 points and usually doesn't give the desired result, but it will have to do.
When the points are too close, we have to revert to Shepards' interpolation which works for horizontal gradients and smaller images.

The syntax is important as it allows a two-stop gradient to finish before the end of the image:

    magick-linear-gradient(
      magick-color-stop(0%,  rgba(0, 0, 0, 0.25)),
      magick-color-stop(50%, rgba(0, 0, 0, 0))
    );